requirements.txt
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
decorator==4.4.2
df2gspread==1.0.4
google-api-python-client==1.6.7
google-auth==1.21.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.2
jedi==0.17.2
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.7.0
pylint==2.6.0
pyquery==1.4.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
traitlets==5.0.4
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
